+++ src/interrupts/idt.zig
@@ -24,7 +24,7 @@ pub const Entry = packed struct {
     pub fn empty() Entry {
         return Entry{ .isr_low = 0, .selector = 0, .reserved = 0, .attributes = 0x8E, .isr_high = 0 };
     }
-    pub fn init(handler: u32) Entry {
+    pub fn init(handler: usize) Entry {
         return Entry{
             .isr_low = @truncate((handler & 0xFFFF)),
             .selector = 0x8, // index of code segment in gdt
@@ -43,9 +43,9 @@ pub const IDTDescr = extern struct {
     /// 1 - size of idt in bytes
     size: u16,
     /// address of idt
-    offset: u32 align(2),
+    offset: usize align(2),
 
-    pub fn init(size: u16, offset: u32) IDTDescr {
+    pub fn init(size: u16, offset: usize) IDTDescr {
         return IDTDescr{ .size = size, .offset = offset };
     }
 
@@ -60,7 +60,7 @@ pub const IDTDescr = extern struct {
 
 pub fn setHandler(
     idx: u8,
-    func: u32,
+    func: usize,
 ) void {
     idt[idx] = Entry.init(func);
 }
